+++ CHANGELOG.md
@@ -15,12 +15,17 @@ This version will not be backwards compatible with Version 1.x.
 - Requires PHPUnit 8+
 - Added CHANGELOG.md
 - Added `auth_ping.php` for testing the authorization token
+- Started supporting Web API v2.1
+- Added support for Share-Links API resource
+- Added support for multiple permissions for shared links via sdo/bitmask
 
 ### Changed
 - Updated the docs regarding obtaining an authorization token
+- SharedLinkType updated for Web API v2.1
 
 ### Removed
 - Removed `obtain_api_token.sh` script
+- Support for Shared-Link API resource (succeeded by Share-Links API resource)
 
 ## [1.0.1] - 2017-06-23